backend/requirements-pythonanywhere.txt
Django==6.0.2
requests==2.32.3
djangorestframework==3.16.1
django-cors-headers==4.7.0
python-dotenv==1.2.1
psycopg2-binary==2.9.10
Pillow==12.1.1
matplotlib==3.10.8
django-allauth[google]>=65.0.0
